fix(pan-to-infowindow): move clearMap to separate file
