+++ app/javascript/angular/code/services/_update_crowd_map_layer.js
@@ -1,6 +1,6 @@
 import _ from "underscore";
 import constants from "../../../javascript/constants";
-import { clearMap } from "../../../javascript/mapsUtils";
+import { clearMap } from "../../../javascript/clearMap";
 
 export const updateCrowdMapLayer = (
   map,
